course-improvements/vietnamese-course/asr_service/requirements.txt
# Pin recent compatible versions
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai-whisper==20250625
# Optional (faster on GPU): torch
# torch==2.0.1
